fix: apply rate limits when columns are hydrated
When columns are hydrated the (token bucket) rate limits were not
applied. They were applied when rate limiting support was added in #623)
However, when support for memoized functions to be assigned as column
hydrate functions was removed in #756, the call to h.rateLimit(ctx, d)
was lost and rateLimitDelay was always 0.